+++ cmd/hoarctl/shared.go
@@ -6,6 +6,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/monax/hoard/v8"
+
 	cli "github.com/jawher/mow.cli"
 	"github.com/monax/hoard/v8/grant"
 	"github.com/monax/hoard/v8/reference"
@@ -24,10 +26,10 @@ func addIntOpt(cmd *cli.Cmd, arg, desc string, def int) *int {
 	return opt
 }
 
-func validateChunkSize(cs int) {
-	if cs == 0 {
+func validateChunkSize(chunkSize int64) {
+	if chunkSize == 0 {
 		fatalf("Chunk size cannot be 0")
-	} else if cs > grpcLimit {
+	} else if chunkSize > hoard.MaxChunkSize {
 		fatalf("Chunk size cannot be greater than 4Mb")
 	}
 }
@@ -55,7 +57,7 @@ func jsonString(v interface{}) string {
 func readReferences(send func(ref *reference.Ref) error) func(chunk []byte) error {
 	decoder := json.NewDecoder(os.Stdin)
 	return func(chunk []byte) error {
-		refs := new(reference.Refs)
+		refs := new([]*reference.Ref)
 		err := decoder.Decode(refs)
 		if err != nil {
 			return err
@@ -69,7 +71,7 @@ func readReferences(send func(ref *reference.Ref) error) func(chunk []byte) erro
 		return nil
 	}
 }
-func recvReferences(refs *reference.Refs, recv func() (*reference.Ref, error)) func() ([]byte, error) {
+func recvReferences(refs *[]*reference.Ref, recv func() (*reference.Ref, error)) func() ([]byte, error) {
 	return func() ([]byte, error) {
 		ref, err := recv()
 		if err != nil {