+++ grant/grant.go
@@ -3,18 +3,22 @@ package grant
 import (
 	"fmt"
 
+	"github.com/monax/hoard/v8/versions"
+
 	"github.com/monax/hoard/v8/config"
 	"github.com/monax/hoard/v8/reference"
 )
 
-const defaultGrantVersion = 2
-
 // Seal this reference into a Grant as specified by Spec
-func Seal(secret config.SecretsManager, refs reference.Refs, spec *Spec) (*Grant, error) {
-	grt := &Grant{Spec: spec, Version: defaultGrantVersion}
+func Seal(secret config.SecretsManager, refs []*reference.Ref, spec *Spec) (*Grant, error) {
+	grt := &Grant{Spec: spec, Version: versions.LatestGrantVersion}
 
+	var err error
 	if s := spec.GetPlaintext(); s != nil {
-		grt.EncryptedReferences = PlaintextGrant(refs)
+		grt.EncryptedReferences, err = PlaintextGrant(refs)
+		if err != nil {
+			return nil, err
+		}
 	} else if s := spec.GetSymmetric(); s != nil {
 		secret, err := secret.Provider(s.PublicID)
 		if err != nil {
@@ -39,30 +43,21 @@ func Seal(secret config.SecretsManager, refs reference.Refs, spec *Spec) (*Grant
 }
 
 // Unseal a Grant exposing its secret reference
-func Unseal(secret config.SecretsManager, grt *Grant) (reference.Refs, error) {
+func Unseal(secret config.SecretsManager, grt *Grant) ([]*reference.Ref, error) {
+	// invert version switch, deal with old JSON references
 	if s := grt.Spec.GetPlaintext(); s != nil {
-		switch grt.GetVersion() {
-		default:
-			return PlaintextReferenceV2(grt.EncryptedReferences), nil
-		}
+		return PlaintextReference(grt.EncryptedReferences, grt.GetVersion())
 
-	} else if s := grt.Spec.GetSymmetric(); s != nil {
+	}
+	if s := grt.Spec.GetSymmetric(); s != nil {
 		secret, err := secret.Provider(s.PublicID)
 		if err != nil {
 			return nil, err
 		}
-		switch grt.GetVersion() {
-		default:
-			return SymmetricReferenceV2(grt.EncryptedReferences, secret.SecretKey)
-		}
-
-	} else if s := grt.Spec.GetOpenPGP(); s != nil {
-		switch grt.GetVersion() {
-		default:
-			return OpenPGPReferenceV2(grt.EncryptedReferences, secret.OpenPGP)
-		}
-
-	} else {
-		return nil, fmt.Errorf("grant type %v not recognised", s)
+		return SymmetricReference(grt.EncryptedReferences, secret.SecretKey, grt.GetVersion())
 	}
+	if s := grt.Spec.GetOpenPGP(); s != nil {
+		return OpenPGPReference(grt.EncryptedReferences, secret.OpenPGP, grt.GetVersion())
+	}
+	return nil, fmt.Errorf("grant type not recognised")
 }