+++ grant/openpgp_test.go
@@ -4,6 +4,8 @@ import (
 	"io/ioutil"
 	"testing"
 
+	"github.com/monax/hoard/v8/versions"
+
 	"github.com/monax/hoard/v8/config"
 
 	"github.com/stretchr/testify/assert"
@@ -27,27 +29,27 @@ func TestOpenPGPGrant(t *testing.T) {
 	assert.NoError(t, err)
 
 	// Try to read reference from grant
-	ref, err := OpenPGPReferenceV2(grant, &testPGP)
+	ref, err := OpenPGPReference(grant, &testPGP, versions.LatestGrantVersion)
 	assert.NoError(t, err)
-	assert.EqualValues(t, testRefs, ref)
+	assertRefsEqual(t, testRefs, ref)
 
 	// Create grant from private
 	grant, err = OpenPGPGrant(testRefs, string(keyPrivate), &testPGP)
 	assert.NoError(t, err)
 
 	// Try to read reference from grant
-	ref, err = OpenPGPReferenceV2(grant, &testPGP)
+	ref, err = OpenPGPReference(grant, &testPGP, versions.LatestGrantVersion)
 	assert.NoError(t, err)
-	assert.EqualValues(t, testRefs, ref)
+	assertRefsEqual(t, testRefs, ref)
 
 	// Create grant from signer
 	grant, err = OpenPGPGrant(testRefs, "", &testPGP)
 	assert.NoError(t, err)
 
 	// Try to read reference from grant
-	ref, err = OpenPGPReferenceV2(grant, &testPGP)
+	ref, err = OpenPGPReference(grant, &testPGP, versions.LatestGrantVersion)
 	assert.NoError(t, err)
-	assert.EqualValues(t, testRefs, ref)
+	assertRefsEqual(t, testRefs, ref)
 
 	grant, err = OpenPGPGrant(testRefs, string(keyPublic), nil)
 	assert.Errorf(t, err, "hoard is not currently configured to use openpgp")