+++ hoard.go
@@ -33,9 +33,9 @@ type ObjectService interface {
 type GrantService interface {
 	ObjectService
 	// Seal a reference by encrypting it according to a grant spec
-	Seal(refs reference.Refs, spec *grant.Spec) (*grant.Grant, error)
+	Seal(refs []*reference.Ref, spec *grant.Spec) (*grant.Grant, error)
 	// Unseal a grant by decrypting it and returning the reference
-	Unseal(grt *grant.Grant) (reference.Refs, error)
+	Unseal(grt *grant.Grant) ([]*reference.Ref, error)
 }
 
 // This is our top level API object providing library acting as a deterministic
@@ -67,11 +67,11 @@ func (hrd *Hoard) Name() string {
 	return hrd.name
 }
 
-func (hrd *Hoard) Seal(refs reference.Refs, spec *grant.Spec) (*grant.Grant, error) {
+func (hrd *Hoard) Seal(refs []*reference.Ref, spec *grant.Spec) (*grant.Grant, error) {
 	return grant.Seal(hrd.secrets, refs, spec)
 }
 
-func (hrd *Hoard) Unseal(grt *grant.Grant) (reference.Refs, error) {
+func (hrd *Hoard) Unseal(grt *grant.Grant) ([]*reference.Ref, error) {
 	return grant.Unseal(hrd.secrets, grt)
 }
 
@@ -99,7 +99,7 @@ func (hrd *Hoard) Put(data, salt []byte) (*reference.Ref, error) {
 	if err != nil {
 		return nil, err
 	}
-	return reference.New(address, blob.SecretKey, salt), nil
+	return reference.New(address, blob.SecretKey, salt, int64(len(data))), nil
 }
 
 func (hrd *Hoard) Delete(address []byte) error {
@@ -113,7 +113,7 @@ func (hrd *Hoard) Encrypt(data, salt []byte) (*reference.Ref, []byte, error) {
 		return nil, nil, err
 	}
 	address := hrd.store.Address(blob.EncryptedData)
-	return reference.New(address, blob.SecretKey, salt), blob.EncryptedData, nil
+	return reference.New(address, blob.SecretKey, salt, int64(len(data))), blob.EncryptedData, nil
 }
 
 // Decrypt data using reference