+++ CHANGELOG.md
@@ -1,9 +1,27 @@
 # [Monax Hoard](https://github.com/monax/hoard) Changelog
-## [Unreleased]
+## [9.0.0]
+This is a fairly major release and the client APIs change significantly. However Grant v2s are still supported and the protobuf API is backwards-compatible.
+
+### Added
+- [Hoard] Grants store the refs to their chunks in a new LINK ref type that is followed during dereferencing (Get, UnsealGet, Decrypt). Version 4 grants _always_ store a single LINK ref. This LINK ref is guaranteed to be unique to the Grant (and therefore grants are now unique). This means UnsealDelete can be safely called without the risk of deleting data still referenced by other grants. This also means grants are not linear in the number of chunks used to store them which keeps grants in constant size.
+- [Hoard] Refs now store the size of the plaintext data stored behind them. This allows for easier random access and predictable downloads.
+- [Hoard] Added regression test to check grant-to-plaintext compatibility between versions.
+
 ### Changed
-- Move to pure-js @grpc/grpc-js library
-- Expose more usable methods from the client
-- Make default ChunkSize 3 MiB
+- [JS] Move to pure-js @grpc/grpc-js library
+- [JS] Expose more usable methods from the client (breaking)
+- [JS] Support streaming versions of calls taking BytesLike
+- [Hoard] Make default storage ChunkSize 3 MiB
+- [Hoard] Body can be sent in same message as MustPlaintext Header (but Header data will be normalised out into first message on storage and retrieval)
+- [Hoard] Grants now encode their references using Protobuf rather than JSON (backwards compatible with grant V2)
+
+
+### Fixed
+- [Hoard] Blocking read-then-write write-then-read usage
+- [Hoard] Unnecessary copying for streams
+- [Hoard] Encrypt endpoints not chunking
+- [Cloud] Stat now explicitly checks for NotFound error for Exists flag, and throws other errors
+
 
 
 ## [8.2.3] - 2020-03-27
@@ -88,7 +106,7 @@ This release makes some changes to the Hoard protobuf and service that are backw
 
 ## [5.0.1] - 2019-06-20
 ### Fixed
-- JS client - v5, yarn publish
+- JS client - v5, npm publish
 
 
 ## [5.0.0] - 2019-05-24
@@ -209,7 +227,7 @@ This is the first Hoard open source release and includes:
 	- Hoar-Daemon hoard
 	- Hoar-Control hoarctl CLI
 
-[Unreleased]: https://github.com/monax/hoard/compare/v8.2.3...HEAD
+[9.0.0]: https://github.com/monax/hoard/compare/v8.2.3...v9.0.0
 [8.2.3]: https://github.com/monax/hoard/compare/v8.2.2...v8.2.3
 [8.2.2]: https://github.com/monax/hoard/compare/v8.2.1...v8.2.2
 [8.2.1]: https://github.com/monax/hoard/compare/v8.2.0...v8.2.1